Changing fatorial from float to double

+++ src/Fatorial.cpp
@@ -6,12 +6,12 @@
 #include "Fatorial.h"
 
 
-float Fatorial::fatorial(int n) {
+double Fatorial::fatorial(int n) {
     static Fatorial *instance = new Fatorial();
     if (n == 0 || n == 1) return 1;
     if (instance->fat_lookUpTable[n] != -1) return instance->fat_lookUpTable[n];
     std::cout << "Exec" << std::endl;
-    float fat = n * fatorial(n - 1);
+    double fat = n * fatorial(n - 1);
     instance->fat_lookUpTable[n] = fat;
     return fat;
 }
